Device: add async_write & co
This is a wrapper for the libftdi's `ftdi_write_data_submit` function.
It does not integrate with any kind of Rust executors at this point and
is designed only to avoid the pause `ftdi_read_data` imposes on a thread
trying to read data from FTDI continuously while sometimed writing to
it.