Support setting the module detach mode.
The module detach mode, a feature added in libFTDI 1.5, determines
whether any kernel module (on Linux: ftdi_sio) attached to the device
should automatically be detached before connecting and/or reattached
after closing the device.  Without it, this must be handled
separately in ways that are less straight-forward (such as through
sysfs).

The libFTDI default is AUTO_DETACH_SIO_MODULE, so this patch defaults to
that as well.